moving notification service/receiver into a service package
